Directory update and new features
Structured the directory in a better way, added logging to the application and implemented two different environments: dev & prod
