+++ .github/workflows/server-pull-changes.yml
@@ -16,4 +16,4 @@ jobs:
           ssh-private-key: ${{ secrets.SSH_PRIVATE_KEY }}
 
       - name: Pull changes on EC2
-        run: ssh -o StrictHostKeyChecking=no ec2-user@${{ secrets.EC2_HOST }} -p ${{ secrets.EC2_SSH_PORT }} 'cd /home/ec2-user/DiskoverProject && git pull'
+        run: ssh -o StrictHostKeyChecking=no ec2-user@${{ secrets.EC2_HOST }} -p ${{ secrets.EC2_SSH_PORT }} 'cd /home/ec2-user/DiskoverProject && git pull && python .\EC2Scripts\Utils\switch_env.py'
