+++ EC2Scripts/AWS/upload_to_s3.py
@@ -1,8 +1,12 @@
-import logging
+import os
+import sys
+from pathlib import Path
+sys.path.append(f'{str(Path('.').absolute())}/EC2Scripts')
+
 import boto3
 from botocore.exceptions import ClientError
-import sys
-import os
+
+from Utils import flow_logger, build_log_message, build_log_error
 
 
 def upload_file(file_name, bucket, object_name=None):
@@ -13,7 +17,8 @@ def upload_file(file_name, bucket, object_name=None):
     s3_client = boto3.client('s3')
     try:
         response = s3_client.upload_file(file_name, bucket, object_name, ExtraArgs={'ContentType': 'application/pdf'})
+        flow_logger.info(build_log_message(f'Uploaded file: {file_name}, to bucket: {bucket}'))
     except ClientError as e:
-        logging.error(e)
+        flow_logger.error(build_log_error(e))
         return False
     return True
