+++ EC2Scripts/Database/__init__.py
@@ -1,0 +1,2 @@
+from .build_db_script import DiskoverDB
+from .connection_script import client
