+++ EC2Scripts/Database/sales_coll_validator.py
