+++ EC2Scripts/Processes/__init__.py
@@ -1,0 +1,6 @@
+from .fill_report import fill_report
+from .generate_plots import generate_plots
+from .generate_records import get_records
+from .get_outofstock_products import get_outofstock_products
+from .restock_products import restock_products
+from .upload_json import upload_json
