+++ EC2Scripts/Properties/dev-config.properties
@@ -1,0 +1,15 @@
+[logging]
+log.level=DEBUG
+
+[database]
+database.name=dev-DiskoverDB
+
+[buckets]
+bucket.main=dev-diskoveranalytics.live
+bucket.archive=dev-diskover-archive
+bucket.input=dev-diskover-input
+
+[sns]
+sns.arn=arn:aws:sns:eu-central-1:419022735529:S3-Input
+sns.analysis.message.success=analysis-success
+sns.analysis.message.fail=analysis-fail
