+++ EC2Scripts/Properties/prod-config.properties
@@ -1,0 +1,15 @@
+[logging]
+log.level=DEBUG
+
+[database]
+database.name=DiskoverDB
+
+[buckets]
+bucket.main=diskoveranalytics.live
+bucket.archive=diskover-archive
+bucket.input=diskover-input
+
+[sns]
+sns.arn=arn:aws:sns:eu-central-1:419022735529:S3-Input
+sns.analysis.message.success=analysis-success
+sns.analysis.message.fail=analysis-fail
