+++ EC2Scripts/Utils/switch_env.py
@@ -1,0 +1,5 @@
+from dotenv import load_dotenv, find_dotenv
+import os
+
+load_dotenv(find_dotenv())
+os.environ['env'] = 'prod'
