--- EC2Scripts/empty_bucket.py
@@ -1,7 +1,0 @@
-import boto3
-
-def empty_bucket(bucket_name):
-    s3 = boto3.resource('s3')
-
-    bucket = s3.Bucket(bucket_name)
-    bucket.objects.all().delete()
